col/embed_server/requirements.txt
auto-round==0.9.2
autoawq==0.2.9
transformers==4.57.1
accelerate==1.10.1
pillow==10.4.0
requests==2.32.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
